requirements.txt
langchain-community
langchain_core
uvicorn
langchain_groq
fastapi
langserve
requests
python_dotenv
streamlit